fix SonarQube: Boolean expressions should not be gratuitous (Java)

+++ cap-rule/src/main/java/com/iisigroup/cap/rule/service/impl/RuleTbMntServiceImpl.java
@@ -128,7 +128,7 @@ public class RuleTbMntServiceImpl implements RuleTbMntService {
     @Override
     public void deleteById(String oid) {
         DivRlItm rlItm = rlItmDao.find(oid);
-        if (rlItm.getDivRlDtls() != null && !rlItm.getDivRlDtls().isEmpty()) {
+        if (rlItm != null && rlItm.getDivRlDtls() != null && !rlItm.getDivRlDtls().isEmpty()) {
             rlDtlDao.delete(rlItm.getDivRlDtls());
         }
         if (rlItm != null) {
